DjangoRestUNI/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.2
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.49.0
fpdf2==2.7.8
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
ndg-httpsclient==0.5.1
pillow==10.2.0
prompt-toolkit==3.0.43
pusher==3.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
service-identity==24.1.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
